refactor move savingModComponentSlice to root levelt pageEditor/store/ directory
