+++ src/pageEditor/store/editor/pageEditorTypes.ts
@@ -18,7 +18,7 @@
 import { type AuthRootState } from "@/auth/authTypes";
 import { type LogRootState } from "@/components/logViewer/logViewerTypes";
 import { type ModComponentsRootState } from "@/store/extensionsTypes";
-import { type SavingModComponentState } from "@/pageEditor/store/savingModComponent/savingModComponentSlice";
+import { type SavingModComponentState } from "@/pageEditor/store/savingModComponentSlice";
 import { type SettingsRootState } from "@/store/settings/settingsTypes";
 import { type RuntimeRootState } from "@/pageEditor/store/runtime/runtimeSliceTypes";
 import { type StarterBrickType } from "@/types/starterBrickTypes";
